Provide constants for ZERO and ONE
Introduce two BigNumber constants as properties of the constructor:
The number zero and the number one. These constants are particularly
useful as initial values for computations (like calculating a sum or a
product) and as default parameter values in functions. Providing them
as constants makes the respective code more readable and efficient
than the repeated invocation of the constructor.